requirements.txt
# GUI
PySide6
qtawesome

# Core imaging
Pillow
opencv-python
numpy==1.26.4

# OCR / detection
manga-ocr
paddleocr
paddlepaddle
onnxruntime

# Translation backends
requests
llama-cpp-python

# Hugging Face / model download
transformers
huggingface-hub>=0.19.3,<1.0
safetensors

# Inpainting
simple-lama-inpainting
torch

# Japanese NLP for Auto-Glossary / Name Memory
fugashi>=1.3.0
unidic-lite>=1.0.8

# Notes:
# - This file keeps CPU-safe defaults where possible.
# - On Windows, GPU users should usually install CUDA-enabled torch,
#   paddlepaddle-gpu, and a GPU-capable llama-cpp-python build via conda/pip
#   for their own environment instead of relying on these default wheels.
